rfctr: move table modules out of shapes sub-pkg
